requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
fpdf==1.7.2
gunicorn==21.2.0